backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.3
python-docx>=1.1.0
PyPDF2>=3.0.1
together>=1.3.0
httpx>=0.26.0
python-dotenv>=1.0.0
reportlab>=4.0.8
jinja2>=3.1.3
aiofiles>=23.2.1
numpy<2
scipy<1.14
scikit-learn<1.5
sentence-transformers>=2.7.0,<3.0.0
torch>=2.0.0,<2.2